requirements.txt
google-genai
python-dotenv
numpy
chromadb
langchain
langchain-google-genai
langchain-chroma